Reusable component for feedback button
