+++ src/common/components/errorTemplate/ErrorTemplate.tsx
@@ -1,16 +1,11 @@
 import classNames from 'classnames';
 import { css } from 'emotion';
-import {
-  IconAlertCircle,
-  IconAngleRight,
-  IconSpeechbubbleText,
-} from 'hds-react';
+import { IconAlertCircle } from 'hds-react';
 import React from 'react';
-import { useTranslation } from 'react-i18next';
 
 import Container from '../../../domain/app/layout/Container';
 import { useTheme } from '../../../domain/app/theme/Theme';
-import Button from '../button/Button';
+import FeedbackButton from '../feedbackButton/FeedbackButton';
 import styles from './errorTemplate.module.scss';
 
 interface Props {
@@ -20,11 +15,6 @@ interface Props {
 
 const ErrorPage: React.FC<Props> = ({ buttons, text }) => {
   const { theme } = useTheme();
-  const { t } = useTranslation();
-
-  const goToFeedback = () => {
-    window.open(t('common.feedback.url'), '_self');
-  };
 
   return (
     <Container
@@ -34,14 +24,7 @@ const ErrorPage: React.FC<Props> = ({ buttons, text }) => {
         <IconAlertCircle className={styles.icon} />
         <p>{text}</p>
         <div className={styles.buttonsWrapper}>{buttons}</div>
-        <Button
-          iconLeft={<IconSpeechbubbleText />}
-          iconRight={<IconAngleRight />}
-          onClick={goToFeedback}
-          variant="supplementary"
-        >
-          {t('common.feedback.text')}
-        </Button>
+        <FeedbackButton />
       </div>
     </Container>
   );
